requirements.txt
fastapi
uvicorn
pyotp
requests
PyGithub
